Rename derive_debug → impl_debug.
